// src/types/SpaceSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file is generated by core-types-ts on behalf of typeconv, DO NOT EDIT.
 * For more information, see:
 *  - {@link https://github.com/grantila/core-types-ts}
 *  - {@link https://github.com/grantila/typeconv}
 */

export interface SpaceSchema {
    id?: string;
    name?: string;
    /**
     * @default
     *     space
     */
    type?: string;
    discriminator?: "classRoom" | "hotelRoom" | "office" | "building" | "commonArea" | "conferenceRoom" | "lobby" | "gym" | "pool" | "restaurant";
    driver?: "aws-kinesis" | "butler" | "crestron" | "dmp" | "dormakaba" | "dsc" | "ecobee" | "igor" | "inncom" | "kohost-k7" | "kohost-pms" | "lg" | "lirc" | "mews" | "mht" | "paxton" | "pelican-wireless" | "power-shades" | "rebrandly" | "salto" | "salto-irn" | "se" | "sendgrid" | "sonifi" | "stay-n-touch" | "twilio" | "cloudflare-images" | "cloudflare-stream" | "insperia-privacy";
    /** This is the category id */
    category?: string;
    rooms?: string[];
    subGroups?: string[];
    occupied?: boolean;
    /**
     * @default
     *     {
     *       "active": false,
     *       "allowed": false
     *     }
     */
    eco?: {
        active?: boolean;
        activatedAt?: string | {
            [key: string]: any;
        };
        allowed?: boolean;
    };
    features?: string[];
    maximumOccupancy?: number;
    housekeepingStatus?: "clean" | "dirty" | "inspected" | "pickup";
    serviceStatus?: "inService" | "outOfOrder" | "outOfService";
    systemId?: string;
    [key: string]: any;
}